Создан родительский класс UIElement для последующего создания элементов интерфейса
