+++ CMakeLists.txt
@@ -1,8 +1,8 @@
 cmake_minimum_required(VERSION 3.22)
 project(RTetris)
 
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++17")
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++20")
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_REQUIRED ON)
 
 # SFML Fetch
@@ -21,15 +21,13 @@ message("-- Copying assets to the build")
 file(COPY assets DESTINATION ${CMAKE_BINARY_DIR})
 message("-- Copying assets to the build - done")
 
-message("-- Including headers directories")
-include_directories(include)
-message("-- Including headers directories - done")
-
-message("-- Adding src/*.cpp files to build")
+message("-- Adding source files to build")
 file(GLOB_RECURSE SOURCES "src/*.cpp")
 add_executable(RTetris ${SOURCES})
-message("-- Adding src/*.cpp files to build - done")
+message("-- Adding source files to build - done")
+target_compile_options(RTetris PRIVATE -Wall -Wextra -pedantic)
 
+message("-- Linking external libraries")
 target_link_libraries(RTetris
     PUBLIC
     sfml-audio
@@ -37,5 +35,8 @@ target_link_libraries(RTetris
     sfml-window
     sfml-system
 )
+message("-- Linking external libraries - done")
 
-target_compile_options(RTetris PRIVATE -Wall -Wextra -pedantic)
+message("-- Including headers directories")
+target_include_directories(RTetris PUBLIC include)
+message("-- Including headers directories - done")
