+++ src/Utilities/DebugInfo.cpp
@@ -2,33 +2,24 @@
 
 #include <SFML/Graphics.hpp>
 
-#include <Utilities/DebugInfo.h>
+#include "UI/UIElement.h"
+#include "Utilities/DebugInfo.h"
+#include "Utilities/FPSCounter.h"
 
-#define DEBUG_TEXT_FONT_SIZE 20
+#define DEBUG_TEXT_FONT_SIZE 10
 
-FPS::FPS() : clock(sf::Clock())
+DebugInfo::DebugInfo(float x, float y) : UIElement(x, y), fps(FPS()), font(sf::Font())
 {
-    clock.restart();
-}
-
-int FPS::get()
-{
-    int result = 1000000 / (clock.getElapsedTime().asMicroseconds() + 1);
-
-    clock.restart();
-    return result;
-}
-
-DebugInfo::DebugInfo(sf::RenderWindow *window) : fps(FPS()), font(sf::Font())
-{
-    this->window = window;
     font.loadFromFile("assets/Fonts/Arial.ttf");
-
-    text = sf::Text("", font, DEBUG_TEXT_FONT_SIZE);
+    text = sf::Text("undefined", font, DEBUG_TEXT_FONT_SIZE);
 }
 
-void DebugInfo::draw()
+void DebugInfo::update()
 {
-    text.setString(std::to_string(fps.get()));
-    window->draw(text);
+    text.setString("FPS: " + std::to_string(fps.get()));
+}
+
+void DebugInfo::draw(sf::RenderTarget &target, sf::RenderStates states) const
+{
+    target.draw(text, states);
 }
